refactor: improve docs and tests
